work around directory symlinks being treated as files when listed
on some servers, if a directory is a symlink, it is treated as a file
when a directory listing is requested, so that rather than the
contents of the directory being returned, just the entry for the link
is returned.

always chdir into the directory before requesting a listing to ensure
a proper listing is returned.